Move ProblemFormat
