+++ src/cpp/benders/benders_core/include/antares-xpansion/benders/benders_core/SolverIO.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "ProblemFormat.h"
+#include "antares-xpansion/core/ProblemFormat.h"
 #include "antares-xpansion/multisolver_interface/SolverAbstract.h"
 #include "antares-xpansion/multisolver_interface/SolverConfig.h"
