+++ src/cpp/core/CMakeLists.txt
@@ -1,0 +1,39 @@
+# ===========================================================================
+# CMake configuration
+# ===========================================================================
+
+# ===========================================================================
+# External requirements
+# ===========================================================================
+
+find_package(fmt REQUIRED)
+
+# ===========================================================================
+# Targets
+# ===========================================================================
+
+add_library(core)
+add_library(${PROJECT_NAME}::core ALIAS core)
+target_sources(core PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/ProblemFormatStream.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/core/ProblemFormat.h
+        ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/core/ProblemFormatStream.h
+)
+
+target_include_directories(core PUBLIC
+        ${CMAKE_CURRENT_SOURCE_DIR}/include
+)
+
+target_link_libraries(core PUBLIC
+        fmt::fmt
+)
+
+# Install targets
+install(TARGETS core
+        LIBRARY DESTINATION lib
+        ARCHIVE DESTINATION lib
+)
+
+install(DIRECTORY include/antares-xpansion
+        DESTINATION include
+)
